refactor: 回答ドメインのテスト配置を責務ごとに整理する
mod.rs が再エクスポート以外の責務を持たないように、回答投稿者、回答内容、回答設定のテストをそれぞれの定義ファイルへ移動する。
