+++ server/domain/src/form/answer/submitter.rs
@@ -43,3 +43,88 @@ impl AnswerSubmitter {
         self.user
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::user::models::{AnswerSubmissionRestrictionReason, Role, UserId};
+    use uuid::Uuid;
+
+    fn user_id(seed: u128) -> UserId {
+        UserId::from(Uuid::from_u128(seed))
+    }
+
+    fn active_user(name: &str, id: UserId, role: Role) -> ActiveUser {
+        ActiveUser::new(name.to_string(), id, role)
+    }
+
+    #[test]
+    fn answer_submitter_is_created_when_user_has_no_active_restriction() {
+        let user = active_user("user", user_id(1), Role::StandardUser);
+
+        assert!(AnswerSubmitter::try_new(user, None, Utc::now()).is_ok());
+    }
+
+    #[test]
+    fn answer_submitter_rejects_active_restriction() {
+        let now = Utc::now();
+        let user = active_user("user", user_id(1), Role::StandardUser);
+        let restriction = AnswerSubmissionRestriction::new(
+            *user.id(),
+            AnswerSubmissionRestrictionReason::new("spam".to_string().try_into().unwrap()),
+            user_id(2),
+            now,
+            None,
+        )
+        .unwrap();
+
+        let result = AnswerSubmitter::try_new(user, Some(restriction), now);
+
+        assert_eq!(
+            result,
+            Err(DomainError::AnswerSubmissionRestricted {
+                reason: "spam".to_string(),
+                expires_at: None,
+            })
+        );
+    }
+
+    #[test]
+    fn answer_submitter_ignores_expired_restriction() {
+        let now = Utc::now();
+        let user = active_user("user", user_id(1), Role::StandardUser);
+        let restriction = AnswerSubmissionRestriction::new(
+            *user.id(),
+            AnswerSubmissionRestrictionReason::new("spam".to_string().try_into().unwrap()),
+            user_id(2),
+            now - chrono::Duration::hours(2),
+            Some(now - chrono::Duration::hours(1)),
+        )
+        .unwrap();
+
+        assert!(AnswerSubmitter::try_new(user, Some(restriction), now).is_ok());
+    }
+
+    #[test]
+    fn answer_submitter_rejects_restriction_for_different_user() {
+        let now = Utc::now();
+        let user = active_user("user", user_id(1), Role::StandardUser);
+        let restriction = AnswerSubmissionRestriction::new(
+            user_id(2),
+            AnswerSubmissionRestrictionReason::new("spam".to_string().try_into().unwrap()),
+            user_id(3),
+            now,
+            None,
+        )
+        .unwrap();
+
+        let result = AnswerSubmitter::try_new(user, Some(restriction), now);
+
+        assert_eq!(
+            result,
+            Err(DomainError::InvalidEntity {
+                message: "answer submission restriction must belong to the submitter".to_string(),
+            })
+        );
+    }
+}
